:recycle: Change from dtfpt to wpdt
