+++ composer.json
@@ -1,7 +1,7 @@
 {
     "autoload": {
         "psr-4": {
-            "DTFPT\\": "src/"
+            "WPDT\\": "src/"
         }
     },
     "require": {
